+++ ext/node/lib.rs
@@ -365,6 +365,7 @@ pub fn init_polyfill() -> Extension {
   Extension::builder(env!("CARGO_PKG_NAME"))
     .esm(esm_files)
     .esm_entry_point("internal:deno_node/module_all.ts")
+    .dependencies(vec!["deno_io"])
     .ops(vec![
       crypto::op_node_create_hash::decl(),
       crypto::op_node_hash_update::decl(),
