+++ ext/node/polyfills/_process/streams.mjs
@@ -11,7 +11,7 @@ import {
 import { Duplex, Readable, Writable } from "internal:deno_node/stream.ts";
 import { isWindows } from "internal:deno_node/_util/os.ts";
 import { fs as fsConstants } from "internal:deno_node/internal_binding/constants.ts";
-import * as files from "internal:runtime/40_files.js";
+import * as io from "internal:deno_io/12_io.js";
 
 // https://github.com/nodejs/node/blob/00738314828074243c9a52a228ab4c68b04259ef/lib/internal/bootstrap/switches/is_main_thread.js#L41
 export function createWritableStdioStream(writer, name) {
@@ -140,7 +140,7 @@ function _guessStdinType(fd) {
 
 const _read = function (size) {
   const p = Buffer.alloc(size || 16 * 1024);
-  files.stdin?.read(p).then((length) => {
+  io.stdin?.read(p).then((length) => {
     this.push(length === null ? null : p.slice(0, length));
   }, (error) => {
     this.destroy(error);
@@ -151,7 +151,7 @@ const _read = function (size) {
 // https://github.com/nodejs/node/blob/v18.12.1/lib/internal/bootstrap/switches/is_main_thread.js#L189
 /** Create process.stdin */
 export const initStdin = () => {
-  const fd = files.stdin?.rid;
+  const fd = io.stdin?.rid;
   let stdin;
   const stdinType = _guessStdinType(fd);
 
@@ -210,8 +210,8 @@ export const initStdin = () => {
     }
   }
 
-  stdin.on("close", () => files.stdin?.close());
-  stdin.fd = files.stdin?.rid ?? -1;
+  stdin.on("close", () => io.stdin?.close());
+  stdin.fd = io.stdin?.rid ?? -1;
   Object.defineProperty(stdin, "isTTY", {
     enumerable: true,
     configurable: true,
@@ -221,7 +221,7 @@ export const initStdin = () => {
   });
   stdin._isRawMode = false;
   stdin.setRawMode = (enable) => {
-    files.stdin?.setRaw?.(enable);
+    io.stdin?.setRaw?.(enable);
     stdin._isRawMode = enable;
     return stdin;
   };
