+++ runtime/js/40_process.js
@@ -2,7 +2,7 @@
 
 const core = globalThis.Deno.core;
 const ops = core.ops;
-import { FsFile } from "internal:runtime/40_files.js";
+import { FsFile } from "internal:runtime/30_fs.js";
 import { readAll } from "internal:deno_io/12_io.js";
 import { pathFromURL } from "internal:runtime/06_util.js";
 import { assert } from "internal:deno_web/00_infra.js";
